Agora Controller recebe como campo um objeto Board

+++ Rules/Coordinate.java
@@ -1,10 +1,10 @@
 public class Coordinate {
-  byte initial_i;
-  byte initial_j;
-  byte final_i;
-  byte final_j;
+  int initial_i;
+  int initial_j;
+  int final_i;
+  int final_j;
 
-  public Coordinate(byte initial_i, byte initial_j, byte final_i, byte final_j){
+  public Coordinate(int initial_i, int initial_j, int final_i, int final_j){
     this.initial_i = initial_i;
     this.initial_j = initial_j;
     this.final_i = final_i;
